Functions to get attributes from a cubble object

Functions to get attributes from a cubble object

form(data)

leaves(data, stem)

group_vars(data)

groups(data)

Arguments

data
an cubble object

stem
either "time" or "spatial".

Details

Apart from inheriting attributes names, row.names, and class from the underlying tibble, a cubble object will also have additional attributes: groups, meta and form.

groups(), meta(), and form() provide a more convenient way to extract these attributes without %@% from a cubble. group_vars() extracts the variable name of the group, which can often be useful.

If a cubble object is also a tsibble, then tsibble attributes (key, index, index2, interval) are also preserved and can be accessed via the relevant functions in the tsibble package.

Examples

# extract attributes of a cubble object
form(climate_small)

#> [1] "nested"

climate_small %>% leaves(stem = "time") %>% head(5)

#> # A tibble: 5 × 6
#>   station       lat  long elevation name                ts                  
#>   <fct>       <dbl> <dbl>     <dbl> <fct>               <named list>        
#> 1 ASN00001019 -14.3  127.        23 kalumburu           <tibble [2,147 × 1]>
#> 2 ASN00001019 -14.3  127.        23 kalumburu           <tibble [2,147 × 1]>
#> 3 ASN00001019 -14.3  127.        23 kalumburu           <tibble [2,147 × 1]>
#> 4 ASN00001019 -14.3  127.        23 kalumburu           <tibble [2,147 × 1]>
#> 5 ASN00002012 -18.2  128.       422 halls creek airport <tibble [2,191 × 1]>

climate_small %>% stretch() %>% leaves(stem = "spatial") %>% head(5)

#> Using `date` as index variable.

#> # A tibble: 5 × 5
#>   station       lat  long elevation name               
#>   <fct>       <dbl> <dbl>     <dbl> <fct>              
#> 1 ASN00001019 -14.3  127.      23   kalumburu          
#> 2 ASN00002012 -18.2  128.     422   halls creek airport
#> 3 ASN00003003 -17.9  122.       7.4 broome airport     
#> 4 ASN00006011 -24.9  114.       4   carnarvon airport  
#> 5 ASN00009021 -31.9  116.      15.4 perth airport      

groups(climate_small) %>% head(5)

#> # A tibble: 5 × 2
#>   station           .rows
#>   <fct>       <list<int>>
#> 1 ASN00001019         [1]
#> 2 ASN00002012         [1]
#> 3 ASN00003003         [1]
#> 4 ASN00006011         [1]
#> 5 ASN00009021         [1]

group_vars(climate_small)

#> [1] "station"

# print out the attribute names of cubble with a tsibble underlying class
names(attributes(climate_small %>% stretch()))

#> Using `date` as index variable.

#>  [1] "names"     "row.names" "key"       "index"     "index2"    "interval" 
#>  [7] "groups"    "leaves"    "form"      "class"