Switch a cubble object into the nested form

face_spatial() turns a long cubble back into a nest cubble and can be seen as the inverse operation of face_temporal(). The nested cubble identifies each row by key and is suitable for operations whose output doesn't involve a time index.

face_spatial(data)

Arguments

data

a long cubble object

Value

a cubble object in the nested form

Examples

cb_long <- climate_flat %>%
  as_cubble(key = id, index = date, coords = c(long, lat)) %>%
  face_temporal()

cb_long %>%  face_spatial()
#> # cubble:   id [5]: nested form
#> # bbox:     [115.97, -32.94, 133.55, -12.42]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>   id            lat  long  elev name           wmo_id ts                
#>   <chr>       <dbl> <dbl> <dbl> <chr>           <dbl> <list>            
#> 1 ASN00009021 -31.9  116.  15.4 perth airport   94610 <tibble [366 × 4]>
#> 2 ASN00010311 -31.9  117. 179   york            94623 <tibble [366 × 4]>
#> 3 ASN00010614 -32.9  117. 338   narrogin        94627 <tibble [366 × 4]>
#> 4 ASN00014015 -12.4  131.  30.4 darwin airport  94120 <tibble [366 × 4]>
#> 5 ASN00015131 -17.6  134. 220   elliott         94236 <tibble [366 × 4]>