Matching sites from two data sources

The function includes both spatial and temporal matching. The spatial matching is based on the distance and the distance is calculated using the Vincenty formula assuming earth is sphere with a radius of 6371 km. The temporal matching first filters out the n largest increases, determined by temporal_n_highest, in both datasets, constructs an interval of length temporal_window from one dataset and count the number that large increase from the other dataset falls into the interval constructed.

match_sites(
  df1,
  df2,
  crs = sf::st_crs("OGC:CRS84"),
  which = NULL,
  par = 0,
  spatial_n_each = 1,
  spatial_n_group = 4,
  temporal_matching = TRUE,
  temporal_by,
  temporal_match_fn = match_peak,
  temporal_n_highest = 20,
  temporal_window = 5,
  temporal_min_match = 10,
  ...
)

match_spatial(
  df1,
  df2,
  crs = sf::st_crs("OGC:CRS84"),
  which = NULL,
  par = 0,
  spatial_n_each = 1,
  spatial_n_group = 4,
  return_cubble = FALSE
)

match_temporal(
  data,
  data_id,
  match_id = NULL,
  temporal_by,
  return_cubble = FALSE,
  temporal_match_fn = match_peak,
  temporal_n_highest = 30,
  temporal_window = 5,
  ...
)

Arguments

df1, df2

the two cubble objects to match

crs

a crs object from st_crs

which

character; for Cartesian coordinates only: one of Euclidean, Hausdorff or Frechet; for geodetic coordinates, great circle distances are computed; see details

par

for which equal to Hausdorff or Frechet, optionally use a value between 0 and 1 to densify the geometry

spatial_n_each

integer, the number of matched "station" in df2 for each df1 record

spatial_n_group

integer, the number of matched group (pair) return

temporal_matching

logical, whether to match temporally

temporal_by

in the syntax of c("xxx" = "xxx), the variables to match in df1 and df2

temporal_match_fn

character, the function name on how two time series should be matched

temporal_n_highest

numeric, the number of highest peak used for temporal matching in match_peak

temporal_window

The temporal window allowed in match_peak

temporal_min_match

The minimum number of peak matching for temporal matching in match_peak

...

parameters passing to temporal match

return_cubble

logical (default to false), whether to return the cubble object or a matching summary table

data

the resulting object from spatial matching

data_id

variable name that separates df1 and df2

match_id

variable name that groups the same match

Examples

a1 <- match_spatial(climate_aus, river)
#> Use OGC:CRS84 by default for distance calculation...
# turn with different distance calculation:
a2 <- match_spatial(climate_aus, river, which = "Hausdorff")
#> Use OGC:CRS84 by default for distance calculation...
# tune the number of matches in each group
a3 <- match_spatial(climate_aus, river, spatial_n_each = 5, spatial_n_group = 2)
#> Use OGC:CRS84 by default for distance calculation...