Compute missing summary

Compute missing summary

add_missing_prct(data, ...)

add_missing_dscrb(data, cutoff = 0.99)

Arguments

data
a cubble object

...
variables to compute percentage missing (support tidyselect)

cutoff
the threshold value, above which the variable will be described as "almost all missing".

Details

add_missing_prct() computes the percentage of missing for the selected variables

add_missing_dscrb() gives a descriptive label (no missing, partly missing, and almost all missing) for those variables that have missing percentage computed

Examples

climate_small %>% add_missing_prct(prcp:tmin)

#> Using `date` as index variable.

#> Using `date` as index variable.

#> # Cubble: station-wise: nested form
#> # Group:  station [55]
#>    station       lat  long elevation name     ts       prcp_missing tmax_missing
#>    <fct>       <dbl> <dbl>     <dbl> <fct>    <list>          <dbl>        <dbl>
#>  1 ASN00001019 -14.3  127.      23   kalumbu… <tbl_ts…      0.0387       0.00419
#>  2 ASN00002012 -18.2  128.     422   halls c… <tbl_ts…      0.0799       0.0183 
#>  3 ASN00003003 -17.9  122.       7.4 broome … <tbl_ts…      0.00456      0.00228
#>  4 ASN00006011 -24.9  114.       4   carnarv… <tbl_ts…      0.0598       0.0114 
#>  5 ASN00009021 -31.9  116.      15.4 perth a… <tbl_ts…      0            0      
#>  6 ASN00009193 -32.0  116.      43.1 rottnes… <tbl_ts…      0.0119       0.00502
#>  7 ASN00009518 -34.4  115.      13   cape le… <tbl_ts…      0.0105       0.00229
#>  8 ASN00009789 -33.8  122.      25   esperan… <tbl_ts…      0.157        0.00869
#>  9 ASN00010286 -31.6  117.     217.  cunderd… <tbl_ts…      0.0151       0.00183
#> 10 ASN00010917 -32.7  117.     275   wanderi… <tbl_ts…      0.00456      0      
#> # … with 45 more rows, and 1 more variable: tmin_missing <dbl>

climate_small %>% add_missing_prct(prcp:tmin) %>% add_missing_dscrb()

#> Using `date` as index variable.

#> Using `date` as index variable.

#> # Cubble: station-wise: nested form
#> # Group:  station [55]
#>    station       lat  long elevation name     ts       prcp_missing tmax_missing
#>    <fct>       <dbl> <dbl>     <dbl> <fct>    <list>          <dbl>        <dbl>
#>  1 ASN00001019 -14.3  127.      23   kalumbu… <tbl_ts…      0.0387       0.00419
#>  2 ASN00002012 -18.2  128.     422   halls c… <tbl_ts…      0.0799       0.0183 
#>  3 ASN00003003 -17.9  122.       7.4 broome … <tbl_ts…      0.00456      0.00228
#>  4 ASN00006011 -24.9  114.       4   carnarv… <tbl_ts…      0.0598       0.0114 
#>  5 ASN00009021 -31.9  116.      15.4 perth a… <tbl_ts…      0            0      
#>  6 ASN00009193 -32.0  116.      43.1 rottnes… <tbl_ts…      0.0119       0.00502
#>  7 ASN00009518 -34.4  115.      13   cape le… <tbl_ts…      0.0105       0.00229
#>  8 ASN00009789 -33.8  122.      25   esperan… <tbl_ts…      0.157        0.00869
#>  9 ASN00010286 -31.6  117.     217.  cunderd… <tbl_ts…      0.0151       0.00183
#> 10 ASN00010917 -32.7  117.     275   wanderi… <tbl_ts…      0.00456      0      
#> # … with 45 more rows, and 4 more variables: tmin_missing <dbl>,
#> #   prcp_dscrb <chr>, tmax_dscrb <chr>, tmin_dscrb <chr>