Switch a cubble object into a long form

Switch a cubble object into a long form

stretch(data, col, ...)

Arguments

data
a cubble object

col
the list column to be stretched

...
variables to be stretched into a single column

Details

stretch() switch a cubble object into the long form where the combination of group identifier and timestamp defines a row. The long form cubble is always of class cubble_df and grouped_df.

See also

Examples

climate_flat %>%
  as_cubble(key = id, index = date, coords = c(long, lat)) %>%
  stretch()

#> # cubble:  date, id [5]: long form
#> # bbox:    [115.97, -32.94, 133.55, -12.42]- check gap on long and lat
#> # spatial: lat [dbl], long [dbl], elev [dbl], name [chr], wmo_id [dbl]
#>    id          date        prcp  tmax  tmin
#>    <chr>       <date>     <dbl> <dbl> <dbl>
#>  1 ASN00009021 2020-01-01     0  31.9  15.3
#>  2 ASN00009021 2020-01-02     0  24.9  16.4
#>  3 ASN00009021 2020-01-03     6  23.2  13  
#>  4 ASN00009021 2020-01-04     0  28.4  12.4
#>  5 ASN00009021 2020-01-05     0  35.3  11.6
#>  6 ASN00009021 2020-01-06     0  34.8  13.1
#>  7 ASN00009021 2020-01-07     0  32.8  15.1
#>  8 ASN00009021 2020-01-08     0  30.4  17.4
#>  9 ASN00009021 2020-01-09     0  28.7  17.3
#> 10 ASN00009021 2020-01-10     0  32.6  15.8
#> # … with 1,820 more rows

# another example for stretching two cubbles