Formulate or switch an object into a cubble in the nested form

Create a cubble object or switch an existing cubble object into the nested form

tamp(data, key)

Arguments

data

the data to be converted into a cubble object

key

the spatio identifier. Key can be automatically detected for a cubble object

Details

For an object that is not cubble_df, tamp() will initialise a cubble object upon providing a group variable.

For an existing cubble in the long form, tamp() switches the object back to the nested form. This form is easier to computing group-specific variables.

See also

Examples

# switch to the nested form from the long form
aus_climate %>%
  stretch() %>%
  tamp()
#> # cubble:   id [85]: nested form
#> # bbox:     [115.97, -43.32, 153.13, -12.42]- check gap on long
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long  elev name                   wmo_id ts               
#>    <chr>       <dbl> <dbl> <dbl> <chr>                   <dbl> <list>           
#>  1 ASN00009021 -31.9  116.  15.4 perth airport           94610 <tibble [366 × 4…
#>  2 ASN00010311 -31.9  117. 179   york                    94623 <tibble [366 × 4…
#>  3 ASN00010614 -32.9  117. 338   narrogin                94627 <tibble [366 × 4…
#>  4 ASN00014015 -12.4  131.  30.4 darwin airport          94120 <tibble [366 × 4…
#>  5 ASN00015131 -17.6  134. 220   elliott                 94236 <tibble [366 × 4…
#>  6 ASN00016065 -30.4  137.  76   andamooka               95660 <tibble [366 × 4…
#>  7 ASN00023034 -35.0  139.   2   adelaide airport        94672 <tibble [366 × 4…
#>  8 ASN00023373 -34.5  139. 275   nuriootpa viticultural  94681 <tibble [366 × 4…
#>  9 ASN00024024 -34.4  141.  30.1 loxton research centre  94682 <tibble [366 × 4…
#> 10 ASN00024048 -34.2  141.  31.5 renmark aero            95687 <tibble [366 × 4…
#> # … with 75 more rows