Toy climate data

Daily measure (2020-01-01 to 2020-01-10) on precipitation (prcp), maximum temperature (tmax), and minimum temperature (tmin) for 3 melbourne airport stations. stations is the spatial component, (stations_sf as an sf object), meteo has the temporal component (meteo_ts as a tsibble object), climate_flat has both in a single joined table, and climate_mel is the cubble object. See climate_aus on the full dataset.

stations

stations_sf

meteo

meteo_ts

climate_flat

climate_mel

Format

An object of class tbl_df (inherits from tbl, data.frame) with 3 rows and 6 columns.

An object of class sf (inherits from tbl_df, tbl, data.frame) with 3 rows and 5 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 30 rows and 5 columns.

An object of class tbl_ts (inherits from tbl_df, tbl, data.frame) with 30 rows and 5 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 30 rows and 10 columns.

An object of class spatial_cubble_df (inherits from cubble_df, tbl_df, tbl, data.frame) with 3 rows and 7 columns.

See also

Examples

cb <- make_cubble(
  spatial = stations, temporal = meteo,
  key = id, index = date, coords = c(long, lat)
)
#> Called from: make_cubble(spatial = stations, temporal = meteo, key = id, index = date, 
#>     coords = c(long, lat))
#> debug: key <- enquo(key)
#> debug: index <- enquo(index)
#> debug: coords <- enquo(coords)
#> debug: if (quo_is_missing(key)) {
#>     if (is_tsibble(temporal)) {
#>         key <- key_vars(temporal)
#>     }
#>     else {
#>         cli::cli_abort("Please supply the {.code key} argument,\n                     or a {.code tsibble} object as the temporal compoenent")
#>     }
#> }
#> debug: if (quo_is_missing(index)) {
#>     if (is_tsibble(temporal)) {
#>         index <- index(temporal)
#>     }
#>     else {
#>         cli::cli_abort("Please supply the {.code index} argument,\n                     or a {.code tsibble} object as the temporal compoenent")
#>     }
#> }
#> debug: if (quo_is_missing(coords)) {
#>     if (is_sf(spatial)) {
#>         cc <- sf::st_coordinates(sf::st_centroid(spatial))
#>         colnames(cc) <- if (sf::st_is_longlat(spatial)) 
#>             c("long", "lat")
#>         else c("x", "y")
#>         spatial <- cbind(spatial, cc)
#>         coords <- colnames(cc)
#>     }
#>     else {
#>         cli::cli_abort("Please supply the {.code coords} argument in the syntax of\n      {.code coords = c(LONG, LAT)}, or\n      use an {.code sf} object as the spatial compoenent")
#>     }
#> } else {
#>     coords <- coords2strvec(coords)
#> }
#> debug: coords <- coords2strvec(coords)
#> debug: common_cols <- intersect(names(spatial), names(temporal))
#> debug: if (!is_null(by)) {
#>     if (by %in% names(temporal) && names(by) %in% names(spatial)) {
#>         names(spatial)[names(spatial) == names(by)] <- by
#>     }
#> } else if (length(common_cols) != 0) {
#>     by <- intersect(names(spatial), names(temporal))[1]
#> } else {
#>     cli::cli_abort("No shared column found.\n    Please supply the shared key using the {.code by} argument")
#> }
#> debug: if (length(common_cols) != 0) {
#>     by <- intersect(names(spatial), names(temporal))[1]
#> } else {
#>     cli::cli_abort("No shared column found.\n    Please supply the shared key using the {.code by} argument")
#> }
#> debug: by <- intersect(names(spatial), names(temporal))[1]
#> debug: if (!is.null(potential_match)) {
#>     if (!inherits(potential_match, "key_tbl")) {
#>         cli::cli_abort("The {.field potential_key_match} need to be the result\n      from {.fn check_key}.")
#>     }
#>     tp <- potential_match$potential_pairs$temporal
#>     sp <- potential_match$potential_pairs$spatial
#>     idx <- match(temporal[[key]], tp)
#>     temporal[[key]] <- ifelse(!is.na(idx), sp[idx], temporal[[key]])
#> }
#> debug: slvl <- spatial[[by]]
#> debug: tlvl <- temporal[[by]]
#> debug: only_spatial <- setdiff(slvl, tlvl)
#> debug: only_temporal <- setdiff(tlvl, slvl)
#> debug: has_unmatch <- length(only_temporal) != 0 | length(only_spatial) != 
#>     0
#> debug: if (length(only_spatial) != 0) cli::cli_alert_warning("Some sites in the spatial table don't have temporal information")
#> debug: if (length(only_temporal) != 0) cli::cli_alert_warning("Some sites in the temporal table don't have spatial information")
#> debug: if (has_unmatch) cli::cli_alert_warning("Use {.fn check_key} to check on the unmatched key\n    The cubble is created only with sites having both spatial and\n    temporal information")
#> debug: spatial <- spatial %>% filter(!by %in% only_spatial)
#> debug: if (is_sf(spatial)) {
#>     spatial <- spatial %>% as_tibble() %>% sf::st_as_sf()
#> }
#> debug: if (is_tsibble(temporal)) {
#>     index <- temporal %@% "index"
#> } else {
#>     index <- as_name(index)
#> }
#> debug: index <- as_name(index)
#> debug: temporal <- temporal %>% filter(!by %in% only_temporal) %>% select(as_name(index), 
#>     setdiff(colnames(temporal), as_name(index)))
#> debug: out <- suppressMessages(dplyr::inner_join(spatial, temporal %>% 
#>     nest(ts = -by)))
#> debug: new_spatial_cubble(out, key = by, index = index, coords = coords)
identical(cb, climate_mel)
#> [1] TRUE
cb2 <- climate_flat %>%
  as_cubble(key = id, index = date, coords = c(long, lat))
identical(cb, climate_mel)
#> [1] TRUE