Functions to extract cubble attributes

Functions to extract cubble attributes

form(data)

is_long(data)

is_nested(data)

spatial(data)

key_vars(data)

key_data(data)

coords(data)

coord_x(data)

coord_y(data)

index(data)

Arguments

data

an cubble object

Details

Apart from inheriting attributes names, row.names, and class from the underlying tibble, a cubble has its site identifier: key, temporal identifier, index, and spatial coordinate reference: coords.

If a cubble object is also a tsibble, then tsibble attributes (key, index, index2, interval) are also preserved and can be accessed via the relevant functions in the tsibble package. (NOT FULLY IMPLEMENTED)

Examples

# extract attributes of a cubble object
form(climate_aus)
#> [1] "nested"
spatial(climate_aus) %>% head(5)
#> # A tibble: 5 × 6
#>   id            lat  long  elev name             wmo_id
#>   <chr>       <dbl> <dbl> <dbl> <chr>             <dbl>
#> 1 ASN00001006 -15.5  128.   3.8 wyndham aero      95214
#> 2 ASN00001007 -13.8  126.   6   troughton island  94102
#> 3 ASN00001018 -16.4  126. 546   mount elizabeth   94211
#> 4 ASN00001019 -14.3  127.  23   kalumburu         94100
#> 5 ASN00001020 -14.1  126.  51   truscott          95101
key_data(climate_aus) %>% head(5)
#> # A tibble: 5 × 2
#>   id                .rows
#>   <chr>       <list<int>>
#> 1 ASN00001006         [1]
#> 2 ASN00001007         [1]
#> 3 ASN00001018         [1]
#> 4 ASN00001019         [1]
#> 5 ASN00001020         [1]
key_vars(climate_aus)
#> [1] "id"
index(climate_aus)
#> [1] "date"
coords(climate_aus)
#> [1] "long" "lat" 
coord_x(climate_aus)
#> [1] "long"
coord_y(climate_aus)
#> [1] "lat"