The constructor for the cubble class

The constructor for the cubble class

as_cubble(data, key, index, coords, ...)

# S3 method for list
as_cubble(data, key, index, coords, output = "all", ...)

# S3 method for tbl_df
as_cubble(data, key, index, coords, ...)

# S3 method for rowwise_df
as_cubble(data, key, index, coords, ...)

# S3 method for cubble_df
tbl_sum(data)

is_cubble(data)

cubble(..., key, index, coords)

Arguments

data

the object to be created or tested as cubble

key

the spatial identifier

index

the time identifier

coords

the coordinates that characterise the spatial dimension

...

a list object to create new cubble

output

either "all" or "unmatch", whether to output all or a list of unmatched summary

Examples

# Declaimer: to make the examples easier, here we first `climate_flat` into
# different classes and show how they can be casted into a cubble. This is to
# demonstrate if your data come in one of the classes, it can be directly cast
# into a cubble. By no mean you need to first transform your data into any of
# the following class and then cast it to cubble.

# If the data is in a tibble:
climate_flat |> as_cubble(key = id, index = date, coords = c(long, lat))
#> # cubble:   id [5]: nested form
#> # bbox:     [115.97, -32.94, 133.55, -12.42]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>   id            lat  long  elev name           wmo_id ts                
#>   <chr>       <dbl> <dbl> <dbl> <chr>           <dbl> <list>            
#> 1 ASN00009021 -31.9  116.  15.4 perth airport   94610 <tibble [366 × 4]>
#> 2 ASN00010311 -31.9  117. 179   york            94623 <tibble [366 × 4]>
#> 3 ASN00010614 -32.9  117. 338   narrogin        94627 <tibble [366 × 4]>
#> 4 ASN00014015 -12.4  131.  30.4 darwin airport  94120 <tibble [366 × 4]>
#> 5 ASN00015131 -17.6  134. 220   elliott         94236 <tibble [366 × 4]>

# If the spatial and temporal information are in two separate tables:
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
spatial <- climate_flat |> select(id:wmo_id) |> distinct()
temporal <- climate_flat |> select(id, date: tmin) |> filter(id != "ASN00009021")
as_cubble(data = list(spatial = spatial, temporal = temporal),
          key = id, index = date, coords = c(long, lat))
#> ! Some sites in the spatial table don't have corresponding temporal information
#> ! Use argument `output = "unmatch"` to check on the unmatched key
#> # cubble:   id [5]: nested form
#> # bbox:     [115.97, -32.94, 133.55, -12.42]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>   id            lat  long  elev name           wmo_id ts                
#>   <chr>       <dbl> <dbl> <dbl> <chr>           <dbl> <list>            
#> 1 ASN00009021 -31.9  116.  15.4 perth airport   94610 <tibble [0 × 4]>  
#> 2 ASN00010311 -31.9  117. 179   york            94623 <tibble [366 × 4]>
#> 3 ASN00010614 -32.9  117. 338   narrogin        94627 <tibble [366 × 4]>
#> 4 ASN00014015 -12.4  131.  30.4 darwin airport  94120 <tibble [366 × 4]>
#> 5 ASN00015131 -17.6  134. 220   elliott         94236 <tibble [366 × 4]>

# If the data is already in a rowwise_df:
dt <- climate_flat |>
  tidyr::nest(ts = date:tmin) |>
  dplyr::rowwise()
dt |> as_cubble(key = id, index = date, coords = c(long, lat))
#> # cubble:   id [5]: nested form
#> # bbox:     [115.97, -32.94, 133.55, -12.42]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>   id            lat  long  elev name           wmo_id ts                
#>   <chr>       <dbl> <dbl> <dbl> <chr>           <dbl> <list>            
#> 1 ASN00009021 -31.9  116.  15.4 perth airport   94610 <tibble [366 × 4]>
#> 2 ASN00010311 -31.9  117. 179   york            94623 <tibble [366 × 4]>
#> 3 ASN00010614 -32.9  117. 338   narrogin        94627 <tibble [366 × 4]>
#> 4 ASN00014015 -12.4  131.  30.4 darwin airport  94120 <tibble [366 × 4]>
#> 5 ASN00015131 -17.6  134. 220   elliott         94236 <tibble [366 × 4]>

# If the data is already in a tsibble, only need to supply `coords`
dt <- climate_flat |> tsibble::as_tsibble(key = id, index = date)
dt |> as_cubble(coords = c(long, lat))
#> # cubble:   id [5]: nested form
#> # bbox:     [115.97, -32.94, 133.55, -12.42]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>   id            lat  long  elev name           wmo_id ts                
#>   <chr>       <dbl> <dbl> <dbl> <chr>           <dbl> <list>            
#> 1 ASN00009021 -31.9  116.  15.4 perth airport   94610 <tbl_ts [366 × 4]>
#> 2 ASN00010311 -31.9  117. 179   york            94623 <tbl_ts [366 × 4]>
#> 3 ASN00010614 -32.9  117. 338   narrogin        94627 <tbl_ts [366 × 4]>
#> 4 ASN00014015 -12.4  131.  30.4 darwin airport  94120 <tbl_ts [366 × 4]>
#> 5 ASN00015131 -17.6  134. 220   elliott         94236 <tbl_ts [366 × 4]>

# If the data is in netcdf:
path <- system.file("ncdf/era5-pressure.nc", package = "cubble")
raw <- ncdf4::nc_open(path)
dt <- as_cubble(raw, vars = c("q", "z"))
#> ℹ More than 10,000 keys: only use the first key to test spatial & temporal variables.