Slicing a cubble

Slicing can be useful when the number of site is too large to be all visualised in a single plot. The slicing family in cubble wraps around the dplyr::slice() family to allow slicing from top and bottom, based on a variable, or in random.

# S3 method for cubble_df
slice_head(data, ...)

# S3 method for cubble_df
slice_tail(data, ...)

# S3 method for cubble_df
slice_min(data, ...)

# S3 method for cubble_df
slice_max(data, ...)

# S3 method for cubble_df
slice_sample(data, ...)

slice_nearby(data, coord, buffer = NA, n = NA)

Arguments

data

a cubble object to slice

...

other arguments passed to the dplyr::slice()

coord

the coordinate of used to slice nearby locations

buffer

the buffer added to the coordinate for slicing

n

the number of nearby points to slice, based on distance

Examples

# slice the first 50 stations from the top/ bottom
library(dplyr)
climate_aus %>% slice_head(n = 50)
#> # cubble:   id [50]: nested form
#> # bbox:     [113.53, -32.05, 128.71, -13.75]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long  elev name                wmo_id ts                
#>    <chr>       <dbl> <dbl> <dbl> <chr>                <dbl> <list>            
#>  1 ASN00001006 -15.5  128.   3.8 wyndham aero         95214 <tibble [366 × 4]>
#>  2 ASN00001007 -13.8  126.   6   troughton island     94102 <tibble [366 × 4]>
#>  3 ASN00001018 -16.4  126. 546   mount elizabeth      94211 <tibble [366 × 4]>
#>  4 ASN00001019 -14.3  127.  23   kalumburu            94100 <tibble [366 × 4]>
#>  5 ASN00001020 -14.1  126.  51   truscott             95101 <tibble [366 × 4]>
#>  6 ASN00001025 -15.4  126. 385   doongan              94215 <tibble [366 × 4]>
#>  7 ASN00002012 -18.2  128. 422   halls creek airport  94212 <tibble [366 × 4]>
#>  8 ASN00002032 -17.0  128. 203   warmun               94213 <tibble [366 × 4]>
#>  9 ASN00002056 -15.8  129.  44   kununurra aero       94216 <tibble [366 × 4]>
#> 10 ASN00002064 -16.6  128. 164   argyle aerodrome     94217 <tibble [366 × 4]>
#> # … with 40 more rows
climate_aus %>% slice_tail(n = 50)
#> # cubble:   id [50]: nested form
#> # bbox:     [143.88, -43.66, 148.35, -39.88]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long  elev name                           wmo_id ts      
#>    <chr>       <dbl> <dbl> <dbl> <chr>                           <dbl> <list>  
#>  1 ASN00091107 -41.0  146.  12   wynyard airport                 95957 <tibble>
#>  2 ASN00091126 -41.2  146.   8   devonport airport               95960 <tibble>
#>  3 ASN00091219 -41.2  147. 198.  scottsdale (west minstone road  94972 <tibble>
#>  4 ASN00091223 -40.9  145. 107.  marrawah                        94950 <tibble>
#>  5 ASN00091237 -41.4  147.   5   launceston (ti tree bend)       94969 <tibble>
#>  6 ASN00091245 -40.7  145.  94   cape grim baps                  94954 <tibble>
#>  7 ASN00091259 -41.1  145. 344   luncheon hill (forestry)        95956 <tibble>
#>  8 ASN00091291 -41.4  146. 294.  sheffield school farm           94955 <tibble>
#>  9 ASN00091292 -40.8  145.   7.5 smithton aerodrome              94953 <tibble>
#> 10 ASN00091293 -41.1  147.   3   low head                        95964 <tibble>
#> # … with 40 more rows

# slice based on the max/ min of a variable

climate_aus %>% slice_max(elev, n = 10)
#> # cubble:   id [10]: nested form
#> # bbox:     [146.27, -37.84, 151.68, -30.21]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long  elev name                 wmo_id ts                
#>    <chr>       <dbl> <dbl> <dbl> <chr>                 <dbl> <list>            
#>  1 ASN00071032 -36.5  148. 1957  thredbo aws           95909 <tibble [366 × 4]>
#>  2 ASN00083085 -37.0  147. 1849  mount hotham          94906 <tibble [366 × 4]>
#>  3 ASN00083084 -36.9  147. 1765  falls creek           94903 <tibble [366 × 4]>
#>  4 ASN00070349 -35.5  149. 1760  mount ginini aws      95925 <tibble [366 × 4]>
#>  5 ASN00083024 -37.1  146. 1707  mount buller          94894 <tibble [366 × 4]>
#>  6 ASN00085291 -37.8  146. 1561  mount baw baw         95901 <tibble [366 × 4]>
#>  7 ASN00072161 -35.9  148. 1482. cabramurra smhea aws  95916 <tibble [366 × 4]>
#>  8 ASN00072091 -35.9  148. 1475  cabramurra smhea      95916 <tibble [366 × 4]>
#>  9 ASN00071041 -36.5  148. 1380  thredbo village       95908 <tibble [366 × 4]>
#> 10 ASN00056229 -30.2  152. 1332  guyra hospital        94772 <tibble [366 × 4]>
climate_aus %>% slice_min(lat, n = 10)
#> # cubble:   id [10]: nested form
#> # bbox:     [145.5, -43.66, 148.01, -42.98]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long  elev name                           wmo_id ts      
#>    <chr>       <dbl> <dbl> <dbl> <chr>                           <dbl> <list>  
#>  1 ASN00094041 -43.7  146. 146.  maatsuyker island lighthouse    94962 <tibble>
#>  2 ASN00094010 -43.5  147.  55   cape bruny lighthouse           94967 <tibble>
#>  3 ASN00094198 -43.5  147.  59.7 cape bruny (cape bruny)         95967 <tibble>
#>  4 ASN00094020 -43.3  147.  17   dover                           94961 <tibble>
#>  5 ASN00094155 -43.2  148. 240   tasman island                   95986 <tibble>
#>  6 ASN00094191 -43.2  147. 831   hartz mountain (keoghs pimple)  94977 <tibble>
#>  7 ASN00097024 -43.1  147. 495   warra                           95962 <tibble>
#>  8 ASN00097083 -43.0  146. 408   scotts peak dam                 95958 <tibble>
#>  9 ASN00094220 -43.0  147.  65   grove (research station)        95977 <tibble>
#> 10 ASN00097080 -43.0  146.  34   low rocky point                 95961 <tibble>

# random sample
climate_aus %>% slice_sample(n = 10)
#> # cubble:   id [10]: nested form
#> # bbox:     [117, -42.9, 151.84, -16.38]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long   elev name                 wmo_id ts                
#>    <chr>       <dbl> <dbl>  <dbl> <chr>                 <dbl> <list>            
#>  1 ASN00009500 -35.0  118.    3   albany                94801 <tibble [366 × 4]>
#>  2 ASN00080091 -36.3  145.  105   kyabram dpi           95833 <tibble [366 × 4]>
#>  3 ASN00065070 -32.2  149.  284   dubbo airport aws     95719 <tibble [366 × 4]>
#>  4 ASN00086383 -37.7  145.   83   coldstream            94864 <tibble [366 × 4]>
#>  5 ASN00009542 -33.7  122.  142   esperance aero        95638 <tibble [366 × 4]>
#>  6 ASN00033119 -21.1  149.   30.3 mackay m.o            94367 <tibble [366 × 4]>
#>  7 ASN00094087 -42.9  147. 1260.  mount wellington      95979 <tibble [366 × 4]>
#>  8 ASN00010524 -32.4  117.  250   brookton              94625 <tibble [366 × 4]>
#>  9 ASN00031037 -16.4  146.    2.9 low isles lighthouse  94285 <tibble [366 × 4]>
#> 10 ASN00040922 -26.6  152.  434.  kingaroy airport      94549 <tibble [366 × 4]>

# slice locations within 1 degree of (130E, 25S)
slice_nearby(climate_aus, coord = c(130, -25), buffer = 3)
#> # cubble:   id [6]: nested form
#> # bbox:     [128.3, -26.27, 132.18, -23.27]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>   id            lat  long  elev name                        wmo_id ts      
#>   <chr>       <dbl> <dbl> <dbl> <chr>                        <dbl> <list>  
#> 1 ASN00013017 -25.0  128.  598  giles meteorological office  94461 <tibble>
#> 2 ASN00015511 -25.3  132.  490. curtin springs               94463 <tibble>
#> 3 ASN00015635 -25.2  131.  492  yulara aero                  94462 <tibble>
#> 4 ASN00015652 -24.3  132.  614  watarrka                     94323 <tibble>
#> 5 ASN00015664 -23.3  129.  454. walungurru                   94321 <tibble>
#> 6 ASN00016097 -26.3  132.  703. ernabella (pukatja)          94474 <tibble>

# slice 5 closest location to (130E, 25S)
slice_nearby(climate_aus, coord = c(130, -25), n = 5)
#> # cubble:   id [5]: nested form
#> # bbox:     [128.3, -25.32, 131.76, -23.27]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>   id            lat  long  elev name      wmo_id ts       long_ref lat_ref  dist
#>   <chr>       <dbl> <dbl> <dbl> <chr>      <dbl> <list>      <dbl>   <dbl> <dbl>
#> 1 ASN00015635 -25.2  131.  492  yulara a…  94462 <tibble>      130     -25  100.
#> 2 ASN00013017 -25.0  128.  598  giles me…  94461 <tibble>      130     -25  171.
#> 3 ASN00015652 -24.3  132.  614  watarrka   94323 <tibble>      130     -25  175.
#> 4 ASN00015511 -25.3  132.  490. curtin s…  94463 <tibble>      130     -25  180.
#> 5 ASN00015664 -23.3  129.  454. walungur…  94321 <tibble>      130     -25  202.