Slicing a cubble

Slicing can be useful when the number of site is too large to be all visualised in a single plot. The slicing family in cubble wraps around the dplyr::slice() family to allow slicing from top and bottom, based on a variable, or in random.

# S3 method for cubble_df
slice_head(.data, ..., n, prop, by)

# S3 method for cubble_df
slice_tail(.data, ..., n, prop, by)

# S3 method for cubble_df
slice_min(.data, order_by, ..., n, prop, by, with_ties, na_rm)

# S3 method for cubble_df
slice_max(.data, order_by, ..., n, prop, by, with_ties, na_rm)

# S3 method for cubble_df
slice_sample(.data, order_by, ..., n, prop, by, with_ties, na_rm)

Arguments

.data

a cubble object to slice

..., n, prop, by

other arguments passed to the dplyr::slice()

order_by, with_ties, na_rm

other arguments passed to the dplyr::slice()

Value

a cubble object

Examples

# slice the first 50 stations from the top/ bottom
library(dplyr)
climate_aus |> slice_head(n = 50)
#> # cubble:   key: id [50], index: date, nested form
#> # extent:   [113.5319, -32.0486, 128.71, -13.7542]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long  elev name                wmo_id ts                
#>    <chr>       <dbl> <dbl> <dbl> <chr>                <dbl> <list>            
#>  1 ASN00001006 -15.5  128.   3.8 wyndham aero         95214 <tibble [366 × 4]>
#>  2 ASN00001007 -13.8  126.   6   troughton island     94102 <tibble [366 × 4]>
#>  3 ASN00001018 -16.4  126. 546   mount elizabeth      94211 <tibble [366 × 4]>
#>  4 ASN00001019 -14.3  127.  23   kalumburu            94100 <tibble [366 × 4]>
#>  5 ASN00001020 -14.1  126.  51   truscott             95101 <tibble [366 × 4]>
#>  6 ASN00001025 -15.4  126. 385   doongan              94215 <tibble [366 × 4]>
#>  7 ASN00002012 -18.2  128. 422   halls creek airport  94212 <tibble [366 × 4]>
#>  8 ASN00002032 -17.0  128. 203   warmun               94213 <tibble [366 × 4]>
#>  9 ASN00002056 -15.8  129.  44   kununurra aero       94216 <tibble [366 × 4]>
#> 10 ASN00002064 -16.6  128. 164   argyle aerodrome     94217 <tibble [366 × 4]>
#> # ℹ 40 more rows
climate_aus |> slice_tail(n = 50)
#> # cubble:   key: id [50], index: date, nested form
#> # extent:   [143.8831, -43.6578, 148.3467, -39.88]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long  elev name                           wmo_id ts      
#>    <chr>       <dbl> <dbl> <dbl> <chr>                           <dbl> <list>  
#>  1 ASN00091107 -41.0  146.  12   wynyard airport                 95957 <tibble>
#>  2 ASN00091126 -41.2  146.   8   devonport airport               95960 <tibble>
#>  3 ASN00091219 -41.2  147. 198.  scottsdale (west minstone road  94972 <tibble>
#>  4 ASN00091223 -40.9  145. 107.  marrawah                        94950 <tibble>
#>  5 ASN00091237 -41.4  147.   5   launceston (ti tree bend)       94969 <tibble>
#>  6 ASN00091245 -40.7  145.  94   cape grim baps                  94954 <tibble>
#>  7 ASN00091259 -41.1  145. 344   luncheon hill (forestry)        95956 <tibble>
#>  8 ASN00091291 -41.4  146. 294.  sheffield school farm           94955 <tibble>
#>  9 ASN00091292 -40.8  145.   7.5 smithton aerodrome              94953 <tibble>
#> 10 ASN00091293 -41.1  147.   3   low head                        95964 <tibble>
#> # ℹ 40 more rows

# slice based on the max/ min of a variable

climate_aus |> slice_max(elev, n = 10)
#> # cubble:   key: id [10], index: date, nested form
#> # extent:   [146.2747, -37.8383, 151.6787, -30.2126]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long  elev name                 wmo_id ts                
#>    <chr>       <dbl> <dbl> <dbl> <chr>                 <dbl> <list>            
#>  1 ASN00056229 -30.2  152. 1332  guyra hospital        94772 <tibble [366 × 4]>
#>  2 ASN00070349 -35.5  149. 1760  mount ginini aws      95925 <tibble [366 × 4]>
#>  3 ASN00071032 -36.5  148. 1957  thredbo aws           95909 <tibble [366 × 4]>
#>  4 ASN00071041 -36.5  148. 1380  thredbo village       95908 <tibble [366 × 4]>
#>  5 ASN00072091 -35.9  148. 1475  cabramurra smhea      95916 <tibble [366 × 4]>
#>  6 ASN00072161 -35.9  148. 1482. cabramurra smhea aws  95916 <tibble [366 × 4]>
#>  7 ASN00083024 -37.1  146. 1707  mount buller          94894 <tibble [366 × 4]>
#>  8 ASN00083084 -36.9  147. 1765  falls creek           94903 <tibble [366 × 4]>
#>  9 ASN00083085 -37.0  147. 1849  mount hotham          94906 <tibble [366 × 4]>
#> 10 ASN00085291 -37.8  146. 1561  mount baw baw         95901 <tibble [366 × 4]>
climate_aus |> slice_min(lat, n = 10)
#> # cubble:   key: id [10], index: date, nested form
#> # extent:   [145.5022, -43.6578, 148.0025, -42.985]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long  elev name                           wmo_id ts      
#>    <chr>       <dbl> <dbl> <dbl> <chr>                           <dbl> <list>  
#>  1 ASN00094010 -43.5  147.  55   cape bruny lighthouse           94967 <tibble>
#>  2 ASN00094020 -43.3  147.  17   dover                           94961 <tibble>
#>  3 ASN00094041 -43.7  146. 146.  maatsuyker island lighthouse    94962 <tibble>
#>  4 ASN00094155 -43.2  148. 240   tasman island                   95986 <tibble>
#>  5 ASN00094191 -43.2  147. 831   hartz mountain (keoghs pimple)  94977 <tibble>
#>  6 ASN00094198 -43.5  147.  59.7 cape bruny (cape bruny)         95967 <tibble>
#>  7 ASN00094220 -43.0  147.  65   grove (research station)        95977 <tibble>
#>  8 ASN00097024 -43.1  147. 495   warra                           95962 <tibble>
#>  9 ASN00097080 -43.0  146.  34   low rocky point                 95961 <tibble>
#> 10 ASN00097083 -43.0  146. 408   scotts peak dam                 95958 <tibble>

# random sample
climate_aus |> slice_sample(n = 10)
#> # cubble:   key: id [10], index: date, nested form
#> # extent:   [127.5297, -38.6297, 146.2564, -18.6494]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long  elev name                    wmo_id ts      
#>    <chr>       <dbl> <dbl> <dbl> <chr>                    <dbl> <list>  
#>  1 ASN00011053 -32.2  128.   3.1 red rocks point          94644 <tibble>
#>  2 ASN00016065 -30.4  137.  76   andamooka                95660 <tibble>
#>  3 ASN00025562 -35.4  141. 110.  lameroo (austin plains)  94690 <tibble>
#>  4 ASN00032078 -18.6  146.  11.8 ingham composite         95291 <tibble>
#>  5 ASN00035069 -24.9  146. 395.  tambo post office        94355 <tibble>
#>  6 ASN00081124 -36.0  146. 129.  yarrawonga               94862 <tibble>
#>  7 ASN00085099 -38.6  146.   3   pound creek              94886 <tibble>
#>  8 ASN00087113 -38.0  144.  10.6 avalon airport           94854 <tibble>
#>  9 ASN00088109 -36.9  145. 141.  mangalore airport        94874 <tibble>
#> 10 ASN00090180 -38.5  144.  95   aireys inlet             94846 <tibble>