Slicing a cubble

Slicing can be useful when the number of site is too large to be all visualised in a single plot. The slicing family in cubble wraps around the dplyr::slice() family to allow slicing from top and bottom, based on a variable, or in random.

# S3 method for cubble_df
slice_head(data, ...)

# S3 method for cubble_df
slice_tail(data, ...)

# S3 method for cubble_df
slice_min(data, ...)

# S3 method for cubble_df
slice_max(data, ...)

# S3 method for cubble_df
slice_sample(data, ...)

Arguments

data

a cubble object to slice

...

other arguments passed to the dplyr::slice()

Examples

# slice the first 50 stations from the top/ bottom
library(dplyr)
aus_climate %>% slice_head(n = 50)
#> # cubble:   id [50]: nested form
#> # bbox:     [115.97, -35.1, 153.13, -12.42]- check gap on long
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long  elev name                   wmo_id ts      
#>    <chr>       <dbl> <dbl> <dbl> <chr>                   <dbl> <list>  
#>  1 ASN00009021 -31.9  116.  15.4 perth airport           94610 <tibble>
#>  2 ASN00010311 -31.9  117. 179   york                    94623 <tibble>
#>  3 ASN00010614 -32.9  117. 338   narrogin                94627 <tibble>
#>  4 ASN00014015 -12.4  131.  30.4 darwin airport          94120 <tibble>
#>  5 ASN00015131 -17.6  134. 220   elliott                 94236 <tibble>
#>  6 ASN00016065 -30.4  137.  76   andamooka               95660 <tibble>
#>  7 ASN00023034 -35.0  139.   2   adelaide airport        94672 <tibble>
#>  8 ASN00023373 -34.5  139. 275   nuriootpa viticultural  94681 <tibble>
#>  9 ASN00024024 -34.4  141.  30.1 loxton research centre  94682 <tibble>
#> 10 ASN00024048 -34.2  141.  31.5 renmark aero            95687 <tibble>
#> # … with 40 more rows
aus_climate %>% slice_tail(n = 50)
#> # cubble:   id [50]: nested form
#> # bbox:     [141.33, -43.32, 151.18, -31.33]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long  elev name                           wmo_id ts      
#>    <chr>       <dbl> <dbl> <dbl> <chr>                           <dbl> <list>  
#>  1 ASN00063303 -33.4  149. 947.  orange airport aws              95726 <tibble>
#>  2 ASN00064017 -31.3  149. 643   coonabarabran airport aws       95728 <tibble>
#>  3 ASN00065068 -33.1  148. 323.  parkes airport aws              95717 <tibble>
#>  4 ASN00065103 -33.4  148. 230.  forbes airport aws              94715 <tibble>
#>  5 ASN00066037 -33.9  151.   6   sydney airport amo              94767 <tibble>
#>  6 ASN00066124 -33.8  151.  55   parramatta north (masons drive  94764 <tibble>
#>  7 ASN00066137 -33.9  151.   6.5 bankstown airport aws           94765 <tibble>
#>  8 ASN00066194 -33.9  151.   3   canterbury racecourse aws       94766 <tibble>
#>  9 ASN00067113 -33.7  151.  24.7 penrith lakes aws               94763 <tibble>
#> 10 ASN00068072 -34.9  151. 109   nowra ran air station  aws      94750 <tibble>
#> # … with 40 more rows

# slice based on the max/ min of a variable
aus_climate %>% slice_max(elev, n = 10)
#> # cubble:   id [10]: nested form
#> # bbox:     [147.53, -36.22, 152.72, -30.21]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long  elev name                          wmo_id ts      
#>    <chr>       <dbl> <dbl> <dbl> <chr>                          <dbl> <list>  
#>  1 ASN00070349 -35.5  149. 1760  mount ginini aws               95925 <tibble>
#>  2 ASN00056229 -30.2  152. 1332  guyra hospital                 94772 <tibble>
#>  3 ASN00063292 -33.6  150. 1080  mount boyce aws                94743 <tibble>
#>  4 ASN00056238 -30.5  152. 1079  armidale airport aws           95773 <tibble>
#>  5 ASN00056037 -30.5  152.  987  armidale (tree group nursery)  94773 <tibble>
#>  6 ASN00082139 -36.2  148.  981  hunters hill                   94878 <tibble>
#>  7 ASN00063303 -33.4  149.  947. orange airport aws             95726 <tibble>
#>  8 ASN00063226 -33.5  150.  900  lithgow (cooerwull)            94741 <tibble>
#>  9 ASN00059140 -30.3  153.  746  dorrigo (old coramba rd)       94789 <tibble>
#> 10 ASN00063291 -33.4  150.  744. bathurst airport aws           94729 <tibble>
aus_climate %>% slice_min(lat, n = 10)
#> # cubble:   id [10]: nested form
#> # bbox:     [142.45, -43.32, 147.51, -38.02]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long  elev name                           wmo_id ts      
#>    <chr>       <dbl> <dbl> <dbl> <chr>                           <dbl> <list>  
#>  1 ASN00094020 -43.3  147.  17   dover                           94961 <tibble>
#>  2 ASN00094008 -42.8  148.   4   hobart airport                  94975 <tibble>
#>  3 ASN00091126 -41.2  146.   8   devonport airport               95960 <tibble>
#>  4 ASN00091219 -41.2  147. 198.  scottsdale (west minstone road  94972 <tibble>
#>  5 ASN00091259 -41.1  145. 344   luncheon hill (forestry)        95956 <tibble>
#>  6 ASN00085099 -38.6  146.   3   pound creek                     94886 <tibble>
#>  7 ASN00090180 -38.5  144.  95   aireys inlet                    94846 <tibble>
#>  8 ASN00090186 -38.3  142.  70.8 warrnambool airport ndb         94837 <tibble>
#>  9 ASN00085072 -38.1  147.   4.6 east sale airport               94907 <tibble>
#> 10 ASN00087113 -38.0  144.  10.6 avalon airport                  94854 <tibble>

# random sample
aus_climate %>% slice_sample(n = 10)
#> # cubble:   id [10]: nested form
#> # bbox:     [137.16, -38.63, 152.72, -29.77]
#> # temporal: date [date], prcp [dbl], tmax [dbl], tmin [dbl]
#>    id            lat  long   elev name                     wmo_id ts      
#>    <chr>       <dbl> <dbl>  <dbl> <chr>                     <dbl> <list>  
#>  1 ASN00087113 -38.0  144.   10.6 avalon airport            94854 <tibble>
#>  2 ASN00085072 -38.1  147.    4.6 east sale airport         94907 <tibble>
#>  3 ASN00059140 -30.3  153.  746   dorrigo (old coramba rd)  94789 <tibble>
#>  4 ASN00016065 -30.4  137.   76   andamooka                 95660 <tibble>
#>  5 ASN00056229 -30.2  152. 1332   guyra hospital            94772 <tibble>
#>  6 ASN00070349 -35.5  149. 1760   mount ginini aws          95925 <tibble>
#>  7 ASN00056242 -29.8  151.  582   inverell (raglan st)      95541 <tibble>
#>  8 ASN00062101 -32.6  150.  471   mudgee airport aws        94727 <tibble>
#>  9 ASN00065103 -33.4  148.  230.  forbes airport aws        94715 <tibble>
#> 10 ASN00085099 -38.6  146.    3   pound creek               94886 <tibble>