Slicing a cubble

Slicing can be useful when the number of site is too large to be all visualised in a single plot. The slicing family in cubble wraps around the dplyr::slice() family to allow slicing from top and bottom, based on a variable, or in random.

# S3 method for cubble_df
slice_head(data, ...)

# S3 method for cubble_df
slice_tail(data, ...)

# S3 method for cubble_df
slice_min(data, ...)

# S3 method for cubble_df
slice_max(data, ...)

# S3 method for cubble_df
slice_sample(data, ...)

Arguments

data
a cubble object to slice

...
other arguments passed to the dplyr::slice()

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:cubble’:
#> 
#>     group_vars, groups

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# slice the first 50 stations from the top/ bottom
climate_small %>% slice_head(n = 50)

#> # Cubble: station-wise: nested form
#> # Group:  station [50]
#>    station       lat  long elevation name                ts                  
#>    <fct>       <dbl> <dbl>     <dbl> <fct>               <list>              
#>  1 ASN00001019 -14.3  127.      23   kalumburu           <tbl_ts [2,147 × 4]>
#>  2 ASN00002012 -18.2  128.     422   halls creek airport <tbl_ts [2,191 × 4]>
#>  3 ASN00003003 -17.9  122.       7.4 broome airport      <tbl_ts [2,192 × 4]>
#>  4 ASN00006011 -24.9  114.       4   carnarvon airport   <tbl_ts [2,192 × 4]>
#>  5 ASN00009021 -31.9  116.      15.4 perth airport       <tbl_ts [2,192 × 4]>
#>  6 ASN00009193 -32.0  116.      43.1 rottnest island     <tbl_ts [2,192 × 4]>
#>  7 ASN00009518 -34.4  115.      13   cape leeuwin        <tbl_ts [2,184 × 4]>
#>  8 ASN00009789 -33.8  122.      25   esperance           <tbl_ts [2,187 × 4]>
#>  9 ASN00010286 -31.6  117.     217.  cunderdin airfield  <tbl_ts [2,190 × 4]>
#> 10 ASN00010917 -32.7  117.     275   wandering           <tbl_ts [2,192 × 4]>
#> # … with 40 more rows

climate_small %>% slice_tail(n = 50)

#> # Cubble: station-wise: nested form
#> # Group:  station [50]
#>    station       lat  long elevation name                        ts             
#>    <fct>       <dbl> <dbl>     <dbl> <fct>                       <list>         
#>  1 ASN00009193 -32.0  116.      43.1 rottnest island             <tbl_ts [2,192…
#>  2 ASN00009518 -34.4  115.      13   cape leeuwin                <tbl_ts [2,184…
#>  3 ASN00009789 -33.8  122.      25   esperance                   <tbl_ts [2,187…
#>  4 ASN00010286 -31.6  117.     217.  cunderdin airfield          <tbl_ts [2,190…
#>  5 ASN00010917 -32.7  117.     275   wandering                   <tbl_ts [2,192…
#>  6 ASN00011052 -30.8  128.     159   forrest                     <tbl_ts [2,191…
#>  7 ASN00012038 -30.8  121.     365.  kalgoorlie boulder airport  <tbl_ts [2,187…
#>  8 ASN00013017 -25.0  128.     598   giles meteorological office <tbl_ts [2,192…
#>  9 ASN00014508 -12.3  137.      51.6 gove airport                <tbl_ts [2,191…
#> 10 ASN00015135 -19.6  134.     376.  tennant creek airport       <tbl_ts [2,189…
#> # … with 40 more rows

# slice based on the max/ min of a variable
climate_small %>% slice_max(elevation, n = 10)

#> # Cubble: station-wise: nested form
#> # Group:  station [10]
#>    station       lat  long elevation name                        ts             
#>    <fct>       <dbl> <dbl>     <dbl> <fct>                       <list>         
#>  1 ASN00096033 -41.9  147.     1057  liawenee                    <tbl_ts [2,192…
#>  2 ASN00013017 -25.0  128.      598  giles meteorological office <tbl_ts [2,192…
#>  3 ASN00015590 -23.8  134.      546  alice springs airport       <tbl_ts [2,192…
#>  4 ASN00002012 -18.2  128.      422  halls creek airport         <tbl_ts [2,191…
#>  5 ASN00015135 -19.6  134.      376. tennant creek airport       <tbl_ts [2,189…
#>  6 ASN00012038 -30.8  121.      365. kalgoorlie boulder airport  <tbl_ts [2,187…
#>  7 ASN00042112 -26.7  150.      305. miles constance street      <tbl_ts [2,165…
#>  8 ASN00030124 -18.3  144.      302. georgetown airport          <tbl_ts [2,192…
#>  9 ASN00044021 -26.4  146.      302. charleville aero            <tbl_ts [2,183…
#> 10 ASN00065070 -32.2  149.      284  dubbo airport aws           <tbl_ts [2,187…

climate_small %>% slice_min(lat, n = 10)

#> # Cubble: station-wise: nested form
#> # Group:  station [10]
#>    station       lat  long elevation name                    ts                 
#>    <fct>       <dbl> <dbl>     <dbl> <fct>                   <list>             
#>  1 ASN00094008 -42.8  148.       4   hobart airport          <tbl_ts [2,192 × 4…
#>  2 ASN00096033 -41.9  147.    1057   liawenee                <tbl_ts [2,192 × 4…
#>  3 ASN00091306 -41.7  147.     148   cressy research station <tbl_ts [2,192 × 4…
#>  4 ASN00091293 -41.1  147.       3   low head                <tbl_ts [2,192 × 4…
#>  5 ASN00091245 -40.7  145.      94   cape grim baps          <tbl_ts [2,191 × 4…
#>  6 ASN00090184 -38.4  142.      45.4 cape nelson lighthouse  <tbl_ts [2,178 × 4…
#>  7 ASN00087031 -37.9  145.      20.1 laverton raaf           <tbl_ts [2,192 × 4…
#>  8 ASN00086282 -37.7  145.     113.  melbourne airport       <tbl_ts [2,192 × 4…
#>  9 ASN00022823 -35.8  137.     158   cape borda              <tbl_ts [2,187 × 4…
#> 10 ASN00074258 -35.6  145.      94   deniliquin airport aws  <tbl_ts [2,192 × 4…

# random sample
climate_small %>% slice_sample(n = 10)

#> # Cubble: station-wise: nested form
#> # Group:  station [10]
#>    station       lat  long elevation name                          ts           
#>    <fct>       <dbl> <dbl>     <dbl> <fct>                         <list>       
#>  1 ASN00068151 -35.1  151.      85   jervis bay point perpendicula <tbl_ts [2,1…
#>  2 ASN00027045 -12.7  142.      18   weipa aero                    <tbl_ts [2,1…
#>  3 ASN00012038 -30.8  121.     365.  kalgoorlie boulder airport    <tbl_ts [2,1…
#>  4 ASN00045025 -28.0  144.     131.  thargomindah airport          <tbl_ts [2,1…
#>  5 ASN00048027 -31.5  146.     260   cobar mo                      <tbl_ts [2,1…
#>  6 ASN00074258 -35.6  145.      94   deniliquin airport aws        <tbl_ts [2,1…
#>  7 ASN00090184 -38.4  142.      45.4 cape nelson lighthouse        <tbl_ts [2,1…
#>  8 ASN00040004 -27.6  153.      24.2 amberley amo                  <tbl_ts [2,1…
#>  9 ASN00006011 -24.9  114.       4   carnarvon airport             <tbl_ts [2,1…
#> 10 ASN00091245 -40.7  145.      94   cape grim baps                <tbl_ts [2,1…