Switch a cubble object into a long form

Switch a cubble object into a long form

stretch(data, key)

Arguments

data
a cubble object

key
the spatio identifier. Key can be automatically detected for a cubble object

Details

stretch() switch a cubble object into the long form where the combination of group identifier and timestamp defines a row. The long form cubble is always of class cubble_df and grouped_df.

See also

Examples

climate_flat %>%
  as_cubble(key = station, index = date, coords = c(long, lat)) %>%
  stretch()

#> # Cubble: time-wise: long form
#> # Key:    station [2]
#> # Leaves: station [fct], lat [dbl], long [dbl], elevation [dbl], name [fct]
#>    station     date        prcp  tmax  tmin
#>    <fct>       <date>     <dbl> <dbl> <dbl>
#>  1 ASN00001019 2020-01-01    46  38.6  25.1
#>  2 ASN00001019 2020-01-02     0  38.8  28.1
#>  3 ASN00001019 2020-01-03   266  37.9  23.6
#>  4 ASN00001019 2020-01-04     0  34.3  26.2
#>  5 ASN00001019 2020-01-05    46  35.4  26.7
#>  6 ASN00001019 2020-01-06   760  27.5  24.8
#>  7 ASN00001019 2020-01-07  1168  31.6  23.2
#>  8 ASN00001019 2020-01-08  1178  32.7  24  
#>  9 ASN00001019 2020-01-09    48  34.2  25.1
#> 10 ASN00001019 2020-01-10     0  35.4  26.7
#> # … with 722 more rows