Formulate or switch an object into a cubble list-column form

Create a cubble object or switch an existing cubble object into the list-column form

tamp(data, key)

Arguments

data
the data to be converted into a cubble object

key
the spatio identifier. Key can be automatically detected for a cubble object

Details

For an object that is not cubble_df, tamp() will initialise a cubble object upon providing a group variable.

For an existing cubble in the long form, tamp() switches the object back to the list-column form. This form is easier to computing group-specific variables.

See also

Examples

library(lubridate)

#> 
#> Attaching package: ‘lubridate’

#> The following objects are masked from ‘package:base’:
#> 
#>     date, intersect, setdiff, union

library(dplyr)

# create a cubble object from a tibble
climate_flat %>% tamp(station)

#> # Cubble: station-wise: list-column
#> # Group:  station [2]
#> # Meta:   station [fct], lat [dbl], long [dbl], elevation [dbl], name [fct]
#>   station       lat  long elevation name                ts                
#>   <fct>       <dbl> <dbl>     <dbl> <fct>               <list>            
#> 1 ASN00001019 -14.3  127.        23 kalumburu           <tibble [366 × 4]>
#> 2 ASN00002012 -18.2  128.       422 halls creek airport <tibble [366 × 4]>

# switch to list-column form from the long form
climate_small %>%
  zoom() %>%
  filter(year(date) == 2020) %>%
  tamp()

#> Using `date` as index variable.

#> Using `date` as index variable.

#> Joining, by = "station"

#> # Cubble: station-wise: list-column
#> # Group:  station [55]
#> # Meta:   station [fct], lat [dbl], long [dbl], elevation [dbl], name [fct]
#>    station       lat  long elevation name                ts                
#>    <fct>       <dbl> <dbl>     <dbl> <fct>               <list>            
#>  1 ASN00001019 -14.3  127.      23   kalumburu           <tbl_ts [366 × 4]>
#>  2 ASN00002012 -18.2  128.     422   halls creek airport <tbl_ts [366 × 4]>
#>  3 ASN00003003 -17.9  122.       7.4 broome airport      <tbl_ts [366 × 4]>
#>  4 ASN00006011 -24.9  114.       4   carnarvon airport   <tbl_ts [366 × 4]>
#>  5 ASN00009021 -31.9  116.      15.4 perth airport       <tbl_ts [366 × 4]>
#>  6 ASN00009193 -32.0  116.      43.1 rottnest island     <tbl_ts [366 × 4]>
#>  7 ASN00009518 -34.4  115.      13   cape leeuwin        <tbl_ts [366 × 4]>
#>  8 ASN00009789 -33.8  122.      25   esperance           <tbl_ts [361 × 4]>
#>  9 ASN00010286 -31.6  117.     217.  cunderdin airfield  <tbl_ts [366 × 4]>
#> 10 ASN00010917 -32.7  117.     275   wandering           <tbl_ts [366 × 4]>
#> # … with 45 more rows